# Refactor to use Web Crypto API