--- a/index.js
+++ b/index.js
@@ -1,6 +1,5 @@
 const ethers = require('ethers');
-const cryptoJS = require('crypto-js');
-const { pbkdf2Sync } = require('pbkdf2');
+const crypto = require('crypto').webcrypto;
 
 // must be signed as uint8array bytes, hence .getBytes()
 const SCA_CREATION_PROOF_MESSAGE = ethers.getBytes(ethers.id('Approve HYTOPIA wallet creation'));
@@ -15,21 +14,20 @@ const CHAIN_IDS = {
  * Public
  */
 
-function aesEncryptWallet(wallet, password, salt) {
-  const key = _pbkdf2(password, salt);
-
+async function aesEncryptWallet(wallet, password, salt) {
   try {
-    return _aesEncrypt(JSON.stringify(getWalletCredentials(wallet)), key);
+    const key = await _pbkdf2(password, salt);
+
+    return await _aesEncrypt(JSON.stringify(getWalletCredentials(wallet)), key);
   } catch (error) {
     throw new Error('Wallet could not be encrypted.');
   }
 }
 
-function aesDecryptWallet(ciphertext, password, salt) {
-  const key = _pbkdf2(password, salt);
-
+async function aesDecryptWallet(ciphertext, password, salt) {
   try {
-    const walletCredentials = JSON.parse(_aesDecrypt(ciphertext, key));
+    const key = await _pbkdf2(password, salt);
+    const walletCredentials = JSON.parse(await _aesDecrypt(ciphertext, key));
 
     return _getWallet(walletCredentials);
   } catch (error) {
@@ -38,7 +36,9 @@ function aesDecryptWallet(ciphertext, password, salt) {
 }
 
 function generateRandomSalt() {
-  return cryptoJS.lib.WordArray.random(512 / 8).toString();
+  const array = new Uint8Array(64);
+  crypto.getRandomValues(array);
+  return Array.from(array, byte => byte.toString(16).padStart(2, '0')).join('');
 }
 
 function generateRandomWallet() {
@@ -100,7 +100,7 @@ async function generateUser(username, email, phone, password) {
   const wallet = generateRandomWallet();
   const salt = generateRandomSalt();
   const authorityAddress = wallet.address;
-  const authorityCiphertext = aesEncryptWallet(wallet, password, salt);
+  const authorityCiphertext = await aesEncryptWallet(wallet, password, salt);
   const authorityProofSignature = await generateScaCreationProofSignature(wallet);
 
   return {
@@ -142,30 +142,53 @@ function _generateNonce(nonceBytes = 32) {
   return ethers.hexlify(ethers.randomBytes(nonceBytes));
 }
 
-function _aesEncrypt(plaintext, key) {
-  try {
-    return cryptoJS.AES.encrypt(plaintext, key).toString();
-  } catch (error) {
-    throw new Error('Could not encrypt.');
-  }
+async function _aesEncrypt(plaintext, key) {
+  const keyBuffer = new Uint8Array(key.match(/.{1,2}/g).map(byte => parseInt(byte, 16)));
+  const iv = crypto.getRandomValues(new Uint8Array(12));
+  const plaintextBuffer = new TextEncoder().encode(plaintext);
+
+  const cryptoKey = await crypto.subtle.importKey('raw', keyBuffer, 'AES-GCM', false, ['encrypt']);
+  const ciphertext = await crypto.subtle.encrypt({ name: 'AES-GCM', iv }, cryptoKey, plaintextBuffer);
+
+  return Array.from(new Uint8Array([...iv, ...new Uint8Array(ciphertext)])).map(byte => byte.toString(16).padStart(2, '0')).join('');
 }
 
-function _aesDecrypt(ciphertext, key) {
-  try {
-    return cryptoJS.AES.decrypt(ciphertext, key).toString(cryptoJS.enc.Utf8);
-  } catch (error) {
-    throw new Error('Could not decrypt');
-  }
+async function _aesDecrypt(ciphertext, key) {
+  const keyBuffer = new Uint8Array(key.match(/.{1,2}/g).map(byte => parseInt(byte, 16)));
+  const ciphertextBuffer = new Uint8Array(ciphertext.match(/.{1,2}/g).map(byte => parseInt(byte, 16)));
+
+  const iv = ciphertextBuffer.subarray(0, 12);
+  const actualCiphertext = ciphertextBuffer.subarray(12);
+
+  const cryptoKey = await crypto.subtle.importKey('raw', keyBuffer, 'AES-GCM', false, ['decrypt']);
+  const plaintextBuffer = await crypto.subtle.decrypt({ name: 'AES-GCM', iv }, cryptoKey, actualCiphertext);
+
+  return new TextDecoder().decode(plaintextBuffer);
 }
 
-function _pbkdf2(password, salt) {
-  return pbkdf2Sync(
-    password,
-    salt,
-    600000, // iterations,
-    32, // key length
-    'sha512',
-  ).toString('hex');
+async function _pbkdf2(password, salt) {
+  const keyMaterial = await crypto.subtle.importKey(
+    'raw',
+    new TextEncoder().encode(password),
+    { name: 'PBKDF2' },
+    false,
+    ['deriveBits', 'deriveKey']
+  );
+  const key = await crypto.subtle.deriveKey(
+    {
+      name: 'PBKDF2',
+      salt: new TextEncoder().encode(salt),
+      iterations: 600000,
+      hash: 'SHA-512',
+    },
+    keyMaterial,
+    { name: 'AES-GCM', length: 32 * 8 },
+    true,
+    ['encrypt', 'decrypt']
+  );
+  const keyBuffer = await crypto.subtle.exportKey('raw', key);
+
+  return Array.from(new Uint8Array(keyBuffer)).map(byte => byte.toString(16).padStart(2, '0')).join('');
 }
 
 /*