--- a/package.json
+++ b/package.json
@@ -17,9 +17,7 @@
   },
   "homepage": "https://github.com/hytopiagg/hytopia-crypto-js#readme",
   "dependencies": {
-    "crypto-js": "^4.2.0",
-    "ethers": "^6.7.1",
-    "pbkdf2": "^3.1.2"
+    "ethers": "^6.7.1"
   },
   "devDependencies": {
     "chai": "^4.3.10",