I replaced the `crypto-js` and `pbkdf2` libraries with their equivalent using the Web Crypto API which is MUCH more performant. Here's some initial numbers:

**Old**
Salt Generation: 0.19ms
PBKDF2 Generation: 5294ms

**New**
Salt Generation: 0.1ms
PBKDF2 Generation: 102ms

Even increasing the iterations back to 600,000 it takes around 300ms to run which is still ~2 orders of magnitude faster. This also reduces size by ~50% when bundled for web.

We can probably replace `ethers` with `viem` and get this even faster and slimmer but it's a pretty good improvement already.

All tests pass but I'm not 100% sure about the refactored encrypt/decrypt functions using the IVs. In theory this will be unique per user and we don't need to store it which is considered safe but let me know if you want me to change it.